configuration avec axios intercepteur du Endoint point refresh pour le refreshToken et a gestion d'erreur en cas d'erreur

+++ src/services/Api.js
@@ -24,13 +24,51 @@ api.interceptors.request.use((config) => {
 //
 api.interceptors.response.use(
   (response) => response,
-  (error) => {
-    if (error.response?.status === 401) {
+
+   async (error) => {
+    const originalRequest = error.config;
+
+ async (error) => {
+    if (error.response?.status === 401  &&
+      !originalRequest._retry) {
+        originalRequest._retry = true;  // pas de requet infinie
       // Token expiré → déconnexion automatique
       useAuthStore.getState().logout();
       sessionStorage.clear();
       window.location.href = "/login";
     }
+
+    try {
+        const refreshToken =
+          sessionStorage.getItem("refreshToken") ||
+          localStorage.getItem("refreshToken");
+
+        if (!refreshToken) {
+          throw new Error("No refresh token");
+        }
+
+        const res = await api.post("/auth/refresh",
+          { refreshToken }
+        );
+
+        const newAccessToken = res.data.accessToken;
+
+        // update storage (les deux pour cohérence simple)
+        sessionStorage.setItem("accessToken", newAccessToken);
+        localStorage.setItem("accessToken", newAccessToken);
+
+        // update header + retry request
+        originalRequest.headers.Authorization =
+          `Bearer ${newAccessToken}`;
+
+        return api(originalRequest);
+
+      } catch (err) {
+        sessionStorage.clear();
+        localStorage.clear();
+        window.location.href = "/login";
+      }
+  }
     return Promise.reject(error);
   },
 );
